requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
typing
opencv-python
Pillow
torch
torchvision
tqdm
streamlit
scikit-image
ultralytics
torch>=2.0.0
torchvision>=0.15.0